requirements.txt
APScheduler==3.10.4
PyYAML==6.0.1
cryptography==41.0.7
pywin32==306; sys_platform == 'win32'
